+++ core/banking/checkpoint_test.go
@@ -75,7 +75,7 @@ func testSimpledScheduledTransfer(t *testing.T) {
 			references []string,
 			feeTransfers []*types.Transfer,
 			feeTransfersAccountTypes []types.AccountType,
-		) ([]*types.TransferResponse, error,
+		) ([]*types.LedgerMovement, error,
 		) {
 			t.Run("ensure transfers are correct", func(t *testing.T) {
 				// transfer is done fully instantly, we should have 2 transfer
@@ -131,7 +131,7 @@ func testSimpledScheduledTransfer(t *testing.T) {
 			references []string,
 			feeTransfers []*types.Transfer,
 			feeTransfersAccountTypes []types.AccountType,
-		) ([]*types.TransferResponse, error,
+		) ([]*types.LedgerMovement, error,
 		) {
 			t.Run("ensure transfers are correct", func(t *testing.T) {
 				// transfer is done fully instantly, we should have 2 transfer