+++ protos/data-node/api/v2/trading_data.pb.gw.go
@@ -2508,11 +2508,11 @@ func request_TradingDataService_ObserveEventBus_0(ctx context.Context, marshaler
 	return stream, metadata, nil
 }
 
-func request_TradingDataService_ObserveTransferResponses_0(ctx context.Context, marshaler runtime.Marshaler, client TradingDataServiceClient, req *http.Request, pathParams map[string]string) (TradingDataService_ObserveTransferResponsesClient, runtime.ServerMetadata, error) {
-	var protoReq ObserveTransferResponsesRequest
+func request_TradingDataService_ObserveLedgerMovements_0(ctx context.Context, marshaler runtime.Marshaler, client TradingDataServiceClient, req *http.Request, pathParams map[string]string) (TradingDataService_ObserveLedgerMovementsClient, runtime.ServerMetadata, error) {
+	var protoReq ObserveLedgerMovementsRequest
 	var metadata runtime.ServerMetadata
 
-	stream, err := client.ObserveTransferResponses(ctx, &protoReq)
+	stream, err := client.ObserveLedgerMovements(ctx, &protoReq)
 	if err != nil {
 		return nil, metadata, err
 	}
@@ -3876,7 +3876,7 @@ func RegisterTradingDataServiceHandlerServer(ctx context.Context, mux *runtime.S
 		return
 	})
 
-	mux.Handle("GET", pattern_TradingDataService_ObserveTransferResponses_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
+	mux.Handle("GET", pattern_TradingDataService_ObserveLedgerMovements_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
 		err := status.Error(codes.Unimplemented, "streaming calls are not yet supported in the in-process transport")
 		_, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
 		runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
@@ -5293,23 +5293,23 @@ func RegisterTradingDataServiceHandlerClient(ctx context.Context, mux *runtime.S
 
 	})
 
-	mux.Handle("GET", pattern_TradingDataService_ObserveTransferResponses_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
+	mux.Handle("GET", pattern_TradingDataService_ObserveLedgerMovements_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
 		ctx, cancel := context.WithCancel(req.Context())
 		defer cancel()
 		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
-		rctx, err := runtime.AnnotateContext(ctx, mux, req, "/datanode.api.v2.TradingDataService/ObserveTransferResponses", runtime.WithHTTPPathPattern("/api/v2/stream/transfer/responses"))
+		rctx, err := runtime.AnnotateContext(ctx, mux, req, "/datanode.api.v2.TradingDataService/ObserveLedgerMovements", runtime.WithHTTPPathPattern("/api/v2/stream/ledger/movements"))
 		if err != nil {
 			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
 			return
 		}
-		resp, md, err := request_TradingDataService_ObserveTransferResponses_0(rctx, inboundMarshaler, client, req, pathParams)
+		resp, md, err := request_TradingDataService_ObserveLedgerMovements_0(rctx, inboundMarshaler, client, req, pathParams)
 		ctx = runtime.NewServerMetadataContext(ctx, md)
 		if err != nil {
 			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
 			return
 		}
 
-		forward_TradingDataService_ObserveTransferResponses_0(ctx, mux, outboundMarshaler, w, req, func() (proto.Message, error) { return resp.Recv() }, mux.GetForwardResponseOptions()...)
+		forward_TradingDataService_ObserveLedgerMovements_0(ctx, mux, outboundMarshaler, w, req, func() (proto.Message, error) { return resp.Recv() }, mux.GetForwardResponseOptions()...)
 
 	})
 
@@ -5507,7 +5507,7 @@ var (
 
 	pattern_TradingDataService_ObserveEventBus_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1, 2, 2, 2, 3, 2, 4}, []string{"api", "v2", "stream", "event", "bus"}, ""))
 
-	pattern_TradingDataService_ObserveTransferResponses_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1, 2, 2, 2, 3, 2, 4}, []string{"api", "v2", "stream", "transfer", "responses"}, ""))
+	pattern_TradingDataService_ObserveLedgerMovements_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1, 2, 2, 2, 3, 2, 4}, []string{"api", "v2", "stream", "ledger", "movements"}, ""))
 
 	pattern_TradingDataService_ListKeyRotations_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1, 2, 2, 2, 3, 2, 4}, []string{"api", "v2", "vega", "keys", "rotations"}, ""))
 
@@ -5647,7 +5647,7 @@ var (
 
 	forward_TradingDataService_ObserveEventBus_0 = runtime.ForwardResponseStream
 
-	forward_TradingDataService_ObserveTransferResponses_0 = runtime.ForwardResponseStream
+	forward_TradingDataService_ObserveLedgerMovements_0 = runtime.ForwardResponseStream
 
 	forward_TradingDataService_ListKeyRotations_0 = runtime.ForwardResponseMessage